app/api/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
aiohttp
sqlalchemy
psycopg2-binary
asyncpg
prometheus_client